Fixed tests for moved endpoints
